+++ src/cacheout/__init__.py
@@ -11,6 +11,7 @@ from .__pkg__ import (
 )
 
 from .cache import Cache
+from .fifo import FIFOCache
 from .lfu import LFUCache
 from .lru import LRUCache
 from .mru import MRUCache